fix: Prefer session.expires_at value in session JWT claims
Stytch session JWTs now include the session's `expires_at` value within the session data claim. If
present, prefer using that value to determine the session's expiration. For backward compatibility
with old tokens that don't include it, continue using the value from the "exp" claim.

This also includes a fix for the "exp" claim's value so it's formatted like the other timestamps.