feat: Add WithBaseURI option to run on internal dev API (#68)
API client setup uses the `env` value to set the correct base URI for that environment. This is
great for the production live and test environments (which are the only supported ones!), but that
makes it difficult to use this within a Stytch development environment as a Stytch API developer.

So this adds a new `WithBaseURI` option to override the derived default.