fix!: Drop support for older Go versions
Based on Go's release support policy, 1.17 is no longer maintained as of
1.19's release in August 2022.

In stytch-go v5.7.0, we started using parts of the Go standard library
(specifically `io.ReadAll`) that are only present in Go 1.16 and later.
So we haven't effectively supported Go 1.15 for a while now.

Our tests pass on Go 1.16, so that could be our minimum Go version if we
need to support back that far for some reason. But I would rather only
support maintained language versions when we have to update the `go`
directive.