feat!: Accept init context for starting API client
The B2C version of the API client prepares for locally validating
session JWTs by starting a JWKS client to periodically fetch the
project's JWKS. This does a synchronous initial fetch that we didn't
provide any way to cancel.

Now, stytchapi.NewAPIClient accepts a context argument to control
cancellation of that initial request.

This does not apply to the B2B API client because it does not start a
client like the B2C API does.