doc: Mark Users.GetPending as deprecated (#118)
With the Users Search endpoint, you can get all of your pending users by adding a filter for
`Status: "pending"`. This can replace any and all existing calls to the Get Pending Users
endpoint.

The Search method also supports more filters to further constrain results. This library provides a
SearchAll convenience helper for iterating through all the results of a search.

We'd like to remove the GetPending method from this library to limit our maintenance work and
remove some complexity for customers.

This marks the GetPending method as deprecated. It will be removed in the next major version of
stytch-go.