dev: Remove failing tests
These tests previously passed because `keyfunc.Get` did not check the
HTTP status codes when fetching the JWKS. This changed in keyfunc v1.4.0
where the default behavior is now to only consider 200 OK to be a
success.

These fake API credentials cause the fetch to return error (400 Bad
Request), which then prevents the API client from starting. These tests
were mostly here as examples, but they become awkward if we want them to
pass. So we can rely on the existing examples in the documentation.