feat: Add NewAPIClientWithContext to cancel init
The B2C version of the API client prepares for locally validating
session JWTs by starting a JWKS client to periodically fetch the
project's JWKS. This does a synchronous initial fetch that we didn't
provide any way to cancel.

Now, stytchapi.NewAPIClientWithContext accepts a context argument that
can cancel that initial request. After the client returns, the context
is ignored, so canceling it will have no effect. The JWKS client is
allowed to continue running in the background.

This does not apply to the B2B API client because it does not start a
client like the B2C API does.